* Combined Java 8 and Java 9+ agents into a single agent * Renamed gmu packages for consistency * Removed unnecessary modules * Renamed phosphor-instrument-jigsaw module to phosphor-driver
