+++ integration-tests/src/test/java/edu/columbia/cs/psl/test/phosphor/BasePhosphorTest.java
@@ -1,6 +1,6 @@
 package edu.columbia.cs.psl.test.phosphor;
 
-import edu.columbia.cs.psl.phosphor.PreMain;
+import edu.columbia.cs.psl.phosphor.Phosphor;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
@@ -8,12 +8,12 @@ import org.junit.Before;
 public class BasePhosphorTest {
     @Before
     public void clearErrorFlag() {
-        PreMain.INSTRUMENTATION_EXCEPTION_OCCURRED = false;
+        Phosphor.INSTRUMENTATION_EXCEPTION_OCCURRED = false;
     }
 
     @After
     public void checkForError() {
-		if(PreMain.INSTRUMENTATION_EXCEPTION_OCCURRED) {
+		if(Phosphor.INSTRUMENTATION_EXCEPTION_OCCURRED) {
 			Assert.fail("Instrumentation error occurred");
 		}
     }
