--- phosphor-distribution/pom.xml
@@ -1,51 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <artifactId>phosphor-distribution</artifactId>
-    <packaging>pom</packaging>
-    <description>Distribution of Phosphor main jar for Java 8, plus Jigsaw (Java9+) JVM instrumenter and runtime agent</description>
-    <parent>
-        <groupId>edu.gmu.swe.phosphor</groupId>
-        <artifactId>phosphor-parent</artifactId>
-        <version>0.1.0-SNAPSHOT</version>
-    </parent>
-    <build>
-        <plugins>
-            <plugin>
-                <artifactId>maven-assembly-plugin</artifactId>
-                <version>3.2.0</version>
-                <executions>
-                    <execution>
-                        <id>distro-assembly</id>
-                        <phase>package</phase>
-                        <goals>
-                            <goal>single</goal>
-                        </goals>
-                        <configuration>
-                            <descriptors>
-                                <descriptor>src/assembly/bin.xml</descriptor>
-                            </descriptors>
-                        </configuration>
-                    </execution>
-                </executions>
-            </plugin>
-        </plugins>
-    </build>
-    <dependencies>
-        <dependency>
-            <groupId>edu.gmu.swe.phosphor</groupId>
-            <artifactId>Phosphor</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>edu.gmu.swe.phosphor</groupId>
-            <artifactId>phosphor-instrument-jigsaw</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>edu.gmu.swe.phosphor</groupId>
-            <artifactId>phosphor-jigsaw-javaagent</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-    </dependencies>
-</project>
