+++ phosphor-driver/src/main/java/edu/columbia/cs/psl/phosphor/driver/Patcher.java
@@ -1,0 +1,7 @@
+package edu.columbia.cs.psl.phosphor.driver;
+
+import java.io.IOException;
+
+public interface Patcher {
+    byte[] patch(String classFileName, byte[] classFileBuffer) throws IOException;
+}
