+++ phosphor-driver/src/main/java/module-info.java
@@ -1,0 +1,6 @@
+module edu.columbia.cs.psl.phosphor.driver {
+    exports edu.columbia.cs.psl.phosphor.driver;
+
+    requires jdk.jlink;
+    requires java.instrument;
+}
