--- phosphor-instrument-jigsaw/src/main/java/edu/gmu/swe/phosphor/instrument/Patcher.java
@@ -1,7 +1,0 @@
-package edu.gmu.swe.phosphor.instrument;
-
-import java.io.IOException;
-
-public interface Patcher {
-    byte[] patch(String classFileName, byte[] classFileBuffer) throws IOException;
-}
