--- phosphor-instrument-jigsaw/src/main/java/module-info.java
@@ -1,5 +1,0 @@
-module edu.gmu.swe.phosphor.instrument {
-    exports edu.gmu.swe.phosphor.instrument;
-    requires jdk.jlink;
-    requires java.instrument;
-}
