--- phosphor-jigsaw-javaagent/src/main/java/module-info.java
@@ -1,4 +1,0 @@
-module edu.columbia.cs.psl.jigsaw.phosphor.javaagent {
-    exports edu.columbia.cs.psl.jigsaw.phosphor.javaagent;
-    requires java.instrument;
-}
